Complete new range assignor - all unit tests pass
